# DS-1187 Full-text indexing of right-to-left PDF files