Fixes https://jira.duraspace.org/browse/DS-1187

This fixes right-to-left full-text indexing for Arabic PDF documents. Already in production in multiple sites since DSpace 1.8.x.