@tdonohue 2015-05-21
Thanks for the PR! At a glance, this looks like a very useful fix to me. I'm scheduling this for possible inclusion in 6.0.

One note here is that PDFBox documentation does warn that specifying "setSortByPosition(true)" could have performance implications. So, we'd want to do some further testing here, or possibly make it an optional config (if we notice performance issues).
http://pdfbox.apache.org/docs/1.8.7/javadocs/org/apache/pdfbox/util/PDFTextStripper.html#setSortByPosition%28boolean%29