@saiful-drtc 2015-05-21
Since most our installations use Arabic PDF, we have been using this method for all our builds. But the original patch I has submitted (3 years back on JIRA) had made it optional. It's available here:
https://jira.duraspace.org/secure/attachment/12529/pdffilter-parseRTL.patch