@helix84 2016-01-27
+1, I did performace benchmarks (with this commit cherry-picked on current master) and it shows no performance penalty on LTR PDFs. On the contrary, all the runs with the patch were a bit faster, but this was run in a VM, so take it with a grain of salt.

with patch: mean 1856 s, stddev 30 s
without patch: mean 2173 s, stddev 14 s