@tdonohue 2016-01-27
+1 by code review. Seems like a good fix to our backend PDF filtering process.